requirements.txt
pyaml
git+https://github.com/LambdaLint/lintipy.git@checks#egg=lintipy
